+++ addOns/authhelper/src/main/java/org/zaproxy/addon/authhelper/ExtensionAuthhelper.java
@@ -19,14 +19,19 @@
  */
 package org.zaproxy.addon.authhelper;
 
+import com.bastiaanjansen.otp.HMACAlgorithm;
+import com.bastiaanjansen.otp.TOTPGenerator;
 import java.awt.EventQueue;
 import java.awt.event.KeyEvent;
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
+import java.time.Duration;
+import java.time.Instant;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Stream;
 import javax.swing.ImageIcon;
 import org.apache.commons.configuration.Configuration;
 import org.apache.commons.lang3.StringUtils;
@@ -47,6 +52,9 @@ import org.parosproxy.paros.model.Session;
 import org.parosproxy.paros.network.HttpMessage;
 import org.parosproxy.paros.view.View;
 import org.zaproxy.addon.authhelper.internal.db.TableJdo;
+import org.zaproxy.addon.commonlib.internal.TotpSupport;
+import org.zaproxy.addon.commonlib.internal.TotpSupport.TotpData;
+import org.zaproxy.addon.commonlib.internal.TotpSupport.TotpGenerator;
 import org.zaproxy.addon.pscan.ExtensionPassiveScan2;
 import org.zaproxy.zap.authentication.FormBasedAuthenticationMethodType;
 import org.zaproxy.zap.authentication.JsonBasedAuthenticationMethodType;
@@ -113,6 +121,35 @@ public class ExtensionAuthhelper extends ExtensionAdaptor {
         return EXTENSION_DEPENDENCIES;
     }
 
+    @Override
+    public void init() {
+        List<String> supportedAlgorithms =
+                Stream.of(HMACAlgorithm.values()).map(HMACAlgorithm::name).toList();
+
+        TotpSupport.setTotpGenerator(
+                new TotpGenerator() {
+
+                    @Override
+                    public List<String> getSupportedAlgorithms() {
+                        return supportedAlgorithms;
+                    }
+
+                    @Override
+                    public String generate(TotpData data, Instant when) {
+                        return new TOTPGenerator.Builder(data.secret())
+                                .withHOTPGenerator(
+                                        builder ->
+                                                builder.withAlgorithm(
+                                                                HMACAlgorithm.valueOf(
+                                                                        data.algorithm()))
+                                                        .withPasswordLength(data.digits()))
+                                .withPeriod(Duration.ofSeconds(data.period()))
+                                .build()
+                                .at(when);
+                    }
+                });
+    }
+
     public AuthhelperParam getParam() {
         if (param == null) {
             param = new AuthhelperParam();
@@ -149,6 +186,8 @@ public class ExtensionAuthhelper extends ExtensionAdaptor {
         AuthUtils.disableBrowserAuthentication();
         BrowserBasedAuthenticationMethodType.stopProxies();
         AuthUtils.clean();
+
+        TotpSupport.setTotpGenerator(null);
     }
 
     @Override