+++ addOns/authhelper/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- Use TOTP data defined under user credentials for Client Script and Browser Based Authentication, when available.
 
 ## [0.24.0] - 2025-03-21
 ### Added